disintermediation-detector/requirements.txt
python-dotenv
langchain
tenacity==8.1.0
functions-framework
langchain-google-vertexai
google-cloud-firestore
google-cloud-aiplatform
